+++ pkg/providers/dns/resource_akamai_dns_zone_test.go
@@ -3,6 +3,7 @@ package dns
 import (
 	"net/http"
 	"os"
+	"regexp"
 	"testing"
 
 	"github.com/akamai/AkamaiOPEN-edgegrid-golang/v7/pkg/dns"
@@ -23,19 +24,185 @@ func TestResDnsZone(t *testing.T) {
 	}
 	recordsetsResp := &dns.RecordSetResponse{Recordsets: make([]dns.Recordset, 2, 2)}
 
-	// This test performs a full life-cycle (CRUD) test
-	t.Run("lifecycle test", func(t *testing.T) {
+	t.Run("when group is not provided and there is no group for the user ", func(t *testing.T) {
 		client := &dns.Mock{}
 
+		client.On("ListGroups",
+			mock.Anything,
+			mock.AnythingOfType("dns.ListGroupRequest"),
+		).Return(&dns.ListGroupResponse{}, nil)
+
+		// work around to skip Delete which fails intentionally
+		err := os.Setenv("DNS_ZONE_SKIP_DELETE", "")
+		require.NoError(t, err)
+		defer func() {
+			err = os.Unsetenv("DNS_ZONE_SKIP_DELETE")
+			require.NoError(t, err)
+		}()
+		useClient(client, func() {
+			resource.UnitTest(t, resource.TestCase{
+				ProviderFactories: testAccProviders,
+				Steps: []resource.TestStep{
+					{
+						Config:      testutils.LoadFixtureString(t, "testdata/TestResDnsZone/create_without_group.tf"),
+						ExpectError: regexp.MustCompile("no group found. Please provide the group."),
+					},
+				},
+			})
+		})
+
+		client.AssertExpectations(t)
+	})
+
+	// This test performs a full life-cycle (CRUD) test
+	t.Run("lifecycle test when group is not found and no. of group is 1", func(t *testing.T) {
+		client := &dns.Mock{}
+		groupListResponse := &dns.ListGroupResponse{
+			Groups: []dns.Group{
+				{
+					GroupID:   1,
+					GroupName: "name",
+					ContractIDs: []string{
+						"1", "2",
+					},
+					Permissions: []string{
+						"DELETE", "READ", "WRITE", "ADD",
+					},
+				},
+			},
+		}
+
+		client.On("ListGroups",
+			mock.Anything,
+			mock.AnythingOfType("dns.ListGroupRequest"),
+		).Return(groupListResponse, nil)
+
 		getCall := client.On("GetZone",
-			mock.Anything, // ctx is irrelevant for this test
+			mock.Anything,
 			zone.Zone,
 		).Return(nil, &dns.Error{
 			StatusCode: http.StatusNotFound,
 		})
 
 		client.On("CreateZone",
-			mock.Anything, // ctx is irrelevant for this test
+			mock.Anything,
+			mock.AnythingOfType("*dns.ZoneCreate"),
+			mock.AnythingOfType("dns.ZoneQueryString"),
+			true,
+		).Return(nil).Run(func(args mock.Arguments) {
+			getCall.ReturnArguments = mock.Arguments{zone, nil}
+		})
+
+		client.On("SaveChangelist",
+			mock.Anything,
+			mock.AnythingOfType("*dns.ZoneCreate"),
+		).Return(nil)
+
+		client.On("SubmitChangelist",
+			mock.Anything,
+			mock.AnythingOfType("*dns.ZoneCreate"),
+		).Return(nil)
+
+		client.On("GetRecordsets",
+			mock.Anything,
+			zone.Zone,
+			mock.AnythingOfType("[]dns.RecordsetQueryArgs"),
+		).Return(recordsetsResp, nil)
+
+		dataSourceName := "akamai_dns_zone.test_without_group"
+
+		// work around to skip Delete which fails intentionally
+		err := os.Setenv("DNS_ZONE_SKIP_DELETE", "")
+		require.NoError(t, err)
+		defer func() {
+			err = os.Unsetenv("DNS_ZONE_SKIP_DELETE")
+			require.NoError(t, err)
+		}()
+		useClient(client, func() {
+			resource.UnitTest(t, resource.TestCase{
+				ProviderFactories: testAccProviders,
+				Steps: []resource.TestStep{
+					{
+						Config: testutils.LoadFixtureString(t, "testdata/TestResDnsZone/create_without_group.tf"),
+						Check: resource.ComposeTestCheckFunc(
+							resource.TestCheckResourceAttr(dataSourceName, "zone", "primaryexampleterraform.io"),
+							resource.TestCheckResourceAttr(dataSourceName, "contract", "ctr1"),
+						),
+					},
+				},
+			})
+		})
+
+		client.AssertExpectations(t)
+	})
+
+	t.Run("when group is not provided and no. of group is more than 1 for the user ", func(t *testing.T) {
+		client := &dns.Mock{}
+		groupListResponse := &dns.ListGroupResponse{
+			Groups: []dns.Group{
+				{
+					GroupID:   1,
+					GroupName: "name",
+					ContractIDs: []string{
+						"1", "2",
+					},
+					Permissions: []string{
+						"DELETE", "READ", "WRITE", "ADD",
+					},
+				},
+				{
+					GroupID:   2,
+					GroupName: "name",
+					ContractIDs: []string{
+						"1", "2",
+					},
+					Permissions: []string{
+						"DELETE", "READ", "WRITE", "ADD",
+					},
+				},
+			},
+		}
+
+		client.On("ListGroups",
+			mock.Anything,
+			mock.AnythingOfType("dns.ListGroupRequest"),
+		).Return(groupListResponse, nil)
+
+		// work around to skip Delete which fails intentionally
+		err := os.Setenv("DNS_ZONE_SKIP_DELETE", "")
+		require.NoError(t, err)
+		defer func() {
+			err = os.Unsetenv("DNS_ZONE_SKIP_DELETE")
+			require.NoError(t, err)
+		}()
+		useClient(client, func() {
+			resource.UnitTest(t, resource.TestCase{
+				ProviderFactories: testAccProviders,
+				Steps: []resource.TestStep{
+					{
+						Config:      testutils.LoadFixtureString(t, "testdata/TestResDnsZone/create_without_group.tf"),
+						ExpectError: regexp.MustCompile("group is a required field when there is more than one group present."),
+					},
+				},
+			})
+		})
+
+		client.AssertExpectations(t)
+	})
+
+	// This test performs a full life-cycle (CRUD) test
+	t.Run("lifecycle test with group", func(t *testing.T) {
+		client := &dns.Mock{}
+
+		getCall := client.On("GetZone",
+			mock.Anything,
+			zone.Zone,
+		).Return(nil, &dns.Error{
+			StatusCode: http.StatusNotFound,
+		})
+
+		client.On("CreateZone",
+			mock.Anything,
 			mock.AnythingOfType("*dns.ZoneCreate"),
 			mock.AnythingOfType("dns.ZoneQueryString"),
 			true,
@@ -44,7 +211,7 @@ func TestResDnsZone(t *testing.T) {
 		})
 
 		client.On("UpdateZone",
-			mock.Anything, // ctx is irrelevant for this test
+			mock.Anything,
 			mock.AnythingOfType("*dns.ZoneCreate"),
 			mock.AnythingOfType("dns.ZoneQueryString"),
 		).Return(nil).Run(func(args mock.Arguments) {
@@ -52,17 +219,17 @@ func TestResDnsZone(t *testing.T) {
 		})
 
 		client.On("SaveChangelist",
-			mock.Anything, // ctx is irrelevant for this test
+			mock.Anything,
 			mock.AnythingOfType("*dns.ZoneCreate"),
 		).Return(nil)
 
 		client.On("SubmitChangelist",
-			mock.Anything, // ctx is irrelevant for this test
+			mock.Anything,
 			mock.AnythingOfType("*dns.ZoneCreate"),
 		).Return(nil)
 
 		client.On("GetRecordsets",
-			mock.Anything, // ctx is irrelevant for this test
+			mock.Anything,
 			zone.Zone,
 			mock.AnythingOfType("[]dns.RecordsetQueryArgs"),
 		).Return(recordsetsResp, nil)
@@ -84,6 +251,9 @@ func TestResDnsZone(t *testing.T) {
 						Config: testutils.LoadFixtureString(t, "testdata/TestResDnsZone/create_primary.tf"),
 						Check: resource.ComposeTestCheckFunc(
 							resource.TestCheckResourceAttr(dataSourceName, "zone", "primaryexampleterraform.io"),
+							resource.TestCheckResourceAttr(dataSourceName, "contract", "ctr1"),
+							resource.TestCheckResourceAttr(dataSourceName, "comment", "This is a test primary zone"),
+							resource.TestCheckResourceAttr(dataSourceName, "group", "grp1"),
 						),
 					},
 					{